Update modules.py
tweaking loss

+++ nn/modules.py
@@ -263,9 +263,9 @@ class CombinedLoss(nn.Module):
 
     def _do_calculation(self, pred, target):
         pred = pred.transpose(3, 1)
-        bce = self.bce(pred, target.float())
-        tversky = self.tversky(pred, target.float())
-        focal = self.focal(pred, target.float())
+        bce = self.bce(pred, target)
+        tversky = self.tversky(pred, target)
+        focal = self.focal(pred, target)
         return self.weights['bce'] * bce + self.weights['tversky'] * tversky + self.weights['focal'] * focal
